Major refactoring and cleaning

+++ src/soot/dexpler/instructions/BinopLitInstruction.java
@@ -57,7 +57,7 @@ public class BinopLitInstruction extends TaggedInstruction {
 
         Local source1 = body.getRegisterLocal(source);
 
-        IntConstant constant = IntConstant.v((int)binOpLitInstr.getNarrowLiteral());
+        IntConstant constant = IntConstant.v(binOpLitInstr.getNarrowLiteral());
 
         Value expr = getExpression(source1, constant);
